myvoice/requirements.txt
asgiref==3.9.1
azure-cognitiveservices-speech==1.45.0
azure-core==1.35.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
